requirements.txt
Of course, here is the list of packages and their descriptions in a plain text format, suitable for a `requirements.txt` file with comments.


# A fast web framework
fastapi==0.104.1
# An ASGI server for FastAPI
uvicorn[standard]==0.24.0
# For handling form data in FastAPI
python-multipart==0.0.6
# A modern HTTP client
httpx==0.28.1
# A simple HTTP library
requests==2.31.0
# For writing concurrent code
asyncio
# Another standard HTTP server
uvicorn

# Data & Database Management
# SQL toolkit and ORM
sqlalchemy==2.0.23
# ODBC driver for databases
pyodbc==5.0.1
# Data analysis library
pandas==2.1.3
# Numerical operations
numpy==1.25.2
# Extensions for datetime objects
python-dateutil==2.8.2
# Windows-specific system APIs
pywin32==306

# AI, Machine Learning & NLP
# Hugging Face models library
transformers==4.35.2
# Core PyTorch library
torch==2.1.0
# Classic ML algorithms
scikit-learn==1.3.0
# LLM application framework
llama-index==0.11.15
# OpenAI-compatible embeddings adapter
llama-index-embeddings-openai==0.2.3
# OpenAI-compatible LLM adapter
llama-index-llms-openai==0.2.7
# Groq API SDK
groq==0.11.0

# General Utilities & Automation
# For SMS, voice, and video
twilio==8.10.0
# For loading environment variables
python-dotenv==1.0.0
# Data validation and settings management
pydantic==2.5.0
# Simple job scheduler
schedule==1.2.0
# Provides colored log output
colorlog==6.7.0

# Testing & Development Tools
# Test framework
pytest==7.4.3
# Code formatter
black==23.11.0
# Cryptographic primitives
cryptography==41.0.7

# Reporting & Visualization
# PDF document generation
reportlab==4.0.4
# Data visualization
matplotlib==3.7.0
# Statistical graphics
seaborn==0.12.0